feat: FP8 KV cache as default (50% VRAM savings)
Switch default kv_dtype from BF16 to float8_e4m3fn everywhere:
TinyserveConfig, offload_experts(), load_and_offload(), and
StaticKVCache.__init__/from_model_config.

Keys and values are cast to FP8 on store, back to BF16 on read.
No logic changes — the existing dtype cast path in update() handles it.

To opt out: kv_dtype=torch.bfloat16

+++ tinyserve/offload.py
@@ -54,7 +54,7 @@ class TinyserveConfig:
     cache_bias: float = 0.0
     adaptive_fate: bool = True
     max_seq_len: int = 0
-    kv_dtype: torch.dtype = field(default_factory=lambda: torch.bfloat16)
+    kv_dtype: torch.dtype = field(default_factory=lambda: torch.float8_e4m3fn)
     gpu_memory_utilization: float = 0.90
     fp8: bool = True
     disk_offload: bool = False
@@ -405,7 +405,7 @@ def offload_model(
     cache_bias: float = 0.0,
     adaptive_fate: bool = True,
     max_seq_len: int = 0,
-    kv_dtype: torch.dtype = torch.bfloat16,
+    kv_dtype: torch.dtype = torch.float8_e4m3fn,
     gpu_memory_utilization: float = 0.90,
     attn_implementation: str | AttentionBackend | None = None,
     disk_offload: bool = False,
@@ -699,7 +699,7 @@ def load_and_offload(
     adaptive_fate: bool = True,
     attn_implementation: str | AttentionBackend | None = None,
     max_seq_len: int = 0,
-    kv_dtype: torch.dtype = torch.bfloat16,
+    kv_dtype: torch.dtype = torch.float8_e4m3fn,
     gpu_memory_utilization: float = 0.90,
     disk_offload: bool = False,
     ram_cache_gb: float = 0,
